Zmiana w orderRepository - typ ID na Long. Zmiany wynikajace z tego w OrderEntityTestSuite.

+++ src/main/java/com/kodilla/ecommercee/repository/OrderRepository.java
@@ -8,6 +8,6 @@ import javax.transaction.Transactional;
 
 @Transactional
 @Repository
-public interface OrderRepository extends CrudRepository<OrderEntity, Integer> {
+public interface OrderRepository extends CrudRepository<OrderEntity, Long> {
 
 }
